requirements.txt
customtkinter>=5.2
Pillow>=10.0
pystray>=0.19
pyinstaller>=6.0
